platform/stm32-e407/: tweak rf230bb.c into the build (with issues)
Besides correctness issues, this also interrupts the build (probably some
vpath confusion). Restarting the build makes it complete successfully.

rf230bb is compiled like this:

1) copy it from contiki/cpu/avr/radio/rf230bb to .
   Changing #include "rf230bb.h" to #include ""platform/stm32-e407/rf230bb.h"

2) platform/stm32-e407/rf230bb.h first include out hal.h (which we've
   generate before) and then the original rf230bb.h

With this detour we can avoid playing tricks with the include path.
The same result could be obtained with the -I- option (although
messily) but this would produce a warning on each cpp invocation.
(cpp suggests the use of -iquote, but that doesn't have the effect
of excluding the current directory from the search path.)